Add Klassik Radio connector (#3064)
* Add connector for Klassik Radio

Klassik Radio is a radio station in Germany. It specialises in classical 
music, Film music and Lounge music. The channel is receivable in over 
300 German cities via FM, throughout Germany via cable, and in Europe 
via satellite. It is also worldwide streamed on the internet.

* Fix time before scrobbling issue on horizonte.cl

It is better to delete the parameters .currentTimeSelector and
.durationSelector. The radio reports the duration of the show and not
the duration of the song (e.g. a show lasts 1 or 2 hours, 3600 or 7200
seconds), so all songs will be sent at 240 seconds, the default maximum
time. This affects songs that are less than 4 minutes long (many cases).

* Update horizonte.cl.js