Silence console logs from content scripts by default (#4067)
* Add a global option to enable debug logging

Some folks have asked for an option to disable debug logging from the
extension, to limit the amount of noise generated in the browser console
across every website. This adds an option (which defaults to disabling
the logging), though it isn't hooked up to anything yet.

I've added English translations for the option's UI to
`src/_locales/en/messages.json` since that seems to be the source file
that translations are automatically generated from, according to the
config in `.tx/config`.

* Don't log to console if debug logging is disabled

Avoids logging anything to the end user's browser console if they
haven't enabled debug logging.

This doesn't affect any of the logging to the extension's service worker
console or the console on the extension's options pages, just to the
user's main browser console when they're on a regular web page.

A couple of things that aren't great here:

This new `getGlobalOption()` function has a lot of duplicate code from
`getOption()` above it. I couldn't use `getOption()` as-is because a) we
don't have a connector to pass as the mandatory `connector` option in
this context, and b) `getOption()` calls `debugLog()` if the requested
option cannot be found, so if we called `debugLog()` from `getOption()`
this could cause an infinite loop.

The `Promise<unknown>` return type of `getGlobalOption()` is pretty
terrible - the promise should only ever resolve to `false` or to one of
the possible values of the `GlobalOptions` interface. I tried to make
that work but I have very minimal experience with TypeScript so I
couldn't work out the syntax.

There are no existing tests for `debugLog()` or `getOption()` here so I
haven't added any to cover this behaviour either.

* Use existing `getOption` function

I had previously tried to import and use this existing `getOption`
function here in `debugLog`, but ran into issues with this error being
thrown:

> Uncaught ReferenceError: Cannot access 'OPTIONS' before initialization

I eventually gave up and wrote a new local `getGlobalOption` function,
which got the job done at the cost of duplicating some code.

However I just noticed that `@/core/storage/browser-storage` is being
imported via an async import function at the top of this file, so I had
a go at importing `@/core/storage/options` the same way and it works
just fine! We no longer need the duplicate function I created earlier,
so I've deleted it here.

* Ensure debug log mesages are printed in order

re: https://github.com/web-scrobbler/web-scrobbler/pull/4067#discussion_r1310393413

Introducing an asynchronous check for user preferences within
`debugLog()` in a previous commit meant that the print order of log
messages was no longer guaranteed to be the same as the order in which
each call to `debugLog()` was made. Implementing a queue for log
messages ensures that they will be printed in the correct order.

* Cache value of DEBUG_LOGGING_ENABLED option

re: PR comment https://github.com/web-scrobbler/web-scrobbler/pull/4067#discussion_r1317722578

It seems excessive to check the current value of this option every time
we want to print a log message, so we can instead check it once when
it's first needed and then cache the result. This means that if a user
changes the option then they'll need to refresh the page before it'll
take effect, but that feels like an acceptable trade-off.